requirements.txt
alembic>=1.6.5
click>=8.0.1
Flask>=1.1.2
Flask-Cors>=3.0.10
Flask-Migrate>=2.7.0
Flask-Script>=2.0.6
Flask-SQLAlchemy>=2.5.1
greenlet>=1.1.0
gunicorn>=20.1.0
itsdangerous>=2.0.1
Jinja2>=3.0.1
Mako>=1.1.4
MarkupSafe>=2.0.1
psycopg2-binary>=2.9.1
python-dateutil>=2.8.1
python-editor>=1.0.4
six>=1.16.0
SQLAlchemy>=1.4.18
Werkzeug>=2.0.1
flask-moment>=0.11.0
flask-wtf>=0.14.3
Flask-RESTful>=0.3.7
aniso8601>=6.0.0
astroid>=2.14.2
ecdsa>=0.18.0
future>=0.18.3
isort>=5.12.0
lazy-object-proxy>=1.9.0
mccabe>=0.7.0
pycryptodome<=3.4.0
pylint>=2.16.3
wrapt>=1.15.0
typed-ast>=1.5.4
python-jose>=3.0.0